+++ src/bluecore_models/models/work.py
@@ -14,7 +14,6 @@ from bluecore_models.utils.db import (
     add_bf_classes,
     add_version,
     update_bf_classes,
-    set_jsonld,
 )
 
 
@@ -30,9 +29,6 @@ class Work(ResourceBase):
 
     def __repr__(self):
         return f"<Work {self.uri}>"
-
-
-event.listen(Work.data, "set", set_jsonld, retval=True)
 
 
 @event.listens_for(Work, "after_insert")
